requirements.txt
Django== 5.0.6  
djangorestframework==3.15.2 
pillow==10.3.0 